Moved ProfileStreamAction from stream to actions namespace
